Fix NPE in traceTransitionSupportFailure when all args are null
PlacementTarget.placementIndex() returns primitive int. The ternary
expression `target != null ? target.placementIndex() : (Integer or null)`
resolves to type int via Java numeric promotion, causing auto-unboxing
of the null Integer branch when both target and pending are null — which
is exactly the case on the transition-support RETRY timeout path.

Adding an explicit (Integer) cast forces the ternary result type to
Integer, so null is stored safely without unboxing.

Reproduces on every run: lane completes → transition support starts with
16 gap blocks → player is frozen (pending fix) → 80-tick timeout fires
→ failTransitionSupport(null, null, (PlacementResult)null) → NPE crash.

+++ src/main/java/com/example/mapart/plan/sweep/grounded/GroundedSingleLaneDebugRunner.java
@@ -1733,7 +1733,7 @@ public final class GroundedSingleLaneDebugRunner {
             PendingPlacementVerification pending
     ) {
         BlockPos failingPos = target != null ? target.worldPos() : pending == null ? null : pending.worldPos();
-        Integer failingIndex = target != null ? target.placementIndex() : pending == null ? null : pending.placementIndex();
+        Integer failingIndex = target != null ? (Integer) target.placementIndex() : pending == null ? null : pending.placementIndex();
         Placement failingPlacement = placement;
         if (failingPlacement == null && failingIndex != null) {
             failingPlacement = transitionSupportPlacementsByIndex.get(failingIndex);
